step 1: definition of Optional[A] and signature for getOrElse
Our goal for the first part of the workshop will be to write
an efficient option-like class along with a couple of high-order combinators
that are free of allocation overhead and don't generate closures.

    final class Optional[+A >: Null](val value: A) extends AnyVal {
      def get: A = value
      def isEmpty = value == null
      def getOrElse[B >: A](alt: => B): B = ???
      def map[B >: Null](f: A => B): Optional[B] = ???
      override def toString = if (isEmpty) "<empty>" else s"$value"
    }

Our little project is inspired by Adam Rosenberger's nalloc library
that does exactly that and then more: https://github.com/arosenberger/nalloc.