step 5: -Xprint:typer reveals the big picture, finally making the bug apparent
-Xprint:typer dumps the representation of the entire program being compiled.
Such detalization often makes the output unwieldy, so we recommend to go for showCode
or -Ymacro-debug-lite first, but sometimes it provides illuminating insights.

In this case, putting macro expansions in context helps us realize that
our expansions never use variables x and y that were referred by
the corresponding macro applications. We need to call isEmpty and value
on something: just `isEmpty` and `value` don't make much sense by themselves.

If getOrElse were a regular method, then naked references to isEmpty and
value would get automatically qualified by `this`, but here we need to spell out
the qualifier ourselves. So, what's the analogue of `this` in macros?

[[syntax trees at end of                     typer]] // Test.scala
package <empty> {
  class C extends scala.AnyRef {
    def <init>(): C = {
      C.super.<init>();
      ()
    };
    override def toString: String = "C"
  };
  object Test extends scala.AnyRef {
    def <init>(): Test.type = {
      Test.super.<init>();
      ()
    };
    def main(args: Array[String]): Unit = {
      val x: Optional[C] = new Optional[C](new C());
      val x1: C = (if (<isEmpty: error>)
        new C()
      else
        <value: error>: C);
      scala.this.Predef.println(scala.Tuple2.apply[Optional[C], C](x, x1));
      val y: Optional[Null] = new Optional[Null](null);
      val y1: C = (if (<isEmpty: error>)
        {
          scala.this.Predef.println("hello");
          new C()
        }
      else
        <value: error>: C);
      scala.this.Predef.println(scala.Tuple2.apply[Optional[Null], C](y, y1))
    }
  }
}