step 17: unfortunately this doesn't always work
Not all definitions have fully-qualified names. For instance, it's a known
issue that one can't qualify program elements defined in the empty package.
Definitions that are not static might also pose a problem (sometimes you
can get away with `Foo.this.something`, but that doesn't work for local variables).

    val Optional = q"_root_.Optional"
    q"""
      val $temp = $prefix
      if ($temp.isEmpty) $Optional.ensuringNotNull($alt) else $temp.value
    """

    [error] /Users/xeno_by/Projects/macrology201/core/src/main/scala/Test.scala:6: object Optional is not a member of package <root>
    [error]     println(new Optional(null).getOrElse(null))
    [error]                                         ^
    [error] one error found

In such cases, your best bet would be to work around (in this particular
situation, working around means moving `Optional` to a named package).

Name clash problems akin to ones we've seen are well-known in macrology,
and actually there exists a fully automatic algorithm that can prevent
name clashes between user-written and macro-generated code. We are working
on implementing it for Scala, and that will obviate the need for creating
fresh names and explicitly writing fully-qualified references.