step 20: we've just encountered the owner chain corruption issue
The bug we've witnessed is probably the most annoying gotcha in Scala macros
at the moment, closely followed by the lack of hygiene. It manifests itself
either as a crash in the icode phase (https://issues.scala-lang.org/browse/SI-5797)
or as a crash in the lambdalift phase (https://issues.scala-lang.org/browse/SI-6665).

This stems from a bigger problem with how scalac's typechecker works with
abstract syntax trees, elaborated in detail at Stack Overflow:
http://stackoverflow.com/questions/20936509.

In a nutshell, scalac maintains two different models of the programs
it processes.

The first model contains abstract syntax trees, which represent
all the code being compiled (sans whitespace and comments), consisting
in all the *.scala files passed to the compiler and all the code generated
by macro expansions in those files. Over the course of compilation, these ASTs
are desugared, assigned with types, transformed, then transformed again,
and again, and again, and finally converted into bytecode.

The second model is called a symbol table. It is also a tree, but it only
contains symbols, which represent just the definitions and their signatures,
without any source code for the bodies of those definitions. This second model
is not strictly necessary, but is just an optimization. It turns out that
a lot of things that the compiler has to do don't require full abstract
syntax trees for everything, just the signatures of the definitions, so
not keeping trees around seems to make sense for performance reasons.
One can say that the symbol table is a cache over the abstract syntax tree
representing the program.

As they say, "there are only two hard problems in Computer Science:
cache invalidation and naming things", and the former is illustrated very
vividly by the situation with Scala's symbol table. A lot of bugs in scalac
are caused by inconsistencies between ASTs and the symbol table, and our problem
is no exception.

A handy way of visualizing the symbol table is with -Yshow-syms. Here's
an excerpt from its output on the problematic macro expansion:

[info] [[symbol layout at end of typer]]
...
[info] * object Test
[info] *     constructor Test
[info] *     method main
[info] *         method foo
[info] *             value x
[info] *         value args
[info] *         value temp$macro$1
[info] *         value y
[info] *     value <local Test>

Whereas here's -Yshow-syms of the same program with the macro application
replaced by manually written code:

[info] [[symbol layout at end of typer]]
...
[info] * object Test
[info] *     constructor Test
[info] *     method main
[info] *         method foo
[info] *             value x
[info] *         value args
[info] *         value temp
[info] *             value y
[info] *     value <local Test>

Comparing the printouts reveals the discrepancy which is causing a crash.
In manually written code, the symbol of `val y` lives under `val temp`,
which mirrors the lexical structure of the code, whereas in the macro
expansion `y` is `temp`'s peer.

An annoying thing about the owner chain bugs is that it takes
some effort to trigger them. For instance, if foo's parameter wasn't by-name,
then everything would compile just fine, even though in that case the
owner chain would also be corrupted.