step 23: solution #1: stepping on really thin ice via manual symbol creation
If we can't use the typechecker to fixup the symbol table for us,
let's fix it ourselves. This entails manually putting all the symbols
of definitions from the prefix under the symbol for `val temp = $prefix`.

In order to do that, we need to call Symbol.setOwner (or an equivalent)
for the topmost definitions in prefix. This can be done by manually traversing
the prefix, but that involves some trickiness, so in Scala 2.11 we expose
c.internal.changeOwner that does exactly this.

changeOwner takes:
1) The tree to be processed,
2) The old owner of the symbols we need to fix (that would be
c.internal.enclosingOwner, which stands for the owner of the
macro application),
3) The new owner (the symbol for temp).

But where do we get the symbol for temp? Quasiquotes such as
`q"val temp = ..."` don't create symbols.

One approach to obtaining the new owner is creating it manually
In order to do that, one needs:
1) The symbol's name. That's easy :)
2) The owner for the newly created symbol. In our case that's easy as well:
temp is nested right under the c.internal.enclosingOwner, so temp's owner
should be c.internal.enclosingOwner.
3) The type signature of the symbol. In our case that's again easy, because
we already know the type of prefix. If that weren't the case, we would be
in trouble, because then we'll have to do the typechecking ourselves using
c.typecheck or maybe even c.internal.typingTransform (covered in later commits).

After synthesizing the new owner, we change the owner for prefix, create
the correct definition for temp and return from the macro.

Note that the updated macro uses the valDef factory instead of just
q"val temp = ${changeOwner(prefix...)}" and refers to the newly introduced
val via its symbol, not its name. This is mandatory. Once a symbol for a
definition is created manually, one must never use ordinary trees to operate
on these definitions.