step 24: solution #2: a higher-level approach
If we revisit solution #1, we'll see that at its core it's very simple -
just a single call to changeOwner is what's necessary to fix the crash
we've encountered.

The most tedious aspect of solution #1 is figuring out the new owner.
Quasiquotes create untyped trees, which don't have any symbols assigned
to them, so we have to create those symbols ourselves.

A simpler approach was recently proposed by Jason Zaugg. It is based on
an observation that owner chain corruption issues only affect later
compilation phases, but not the typechecker. Therefore we can delegate owner
creation to the typechecker and then in a subsequent macro expansion
use the created owners to fixup the corresponding trees.