step 25: the final example in part 1: Optional.map
Another important thing that ought to be covered in this workshop is
how to inline stuff using macros. Let's write a zero-overhead higher-order
function for our optional type.

This commit only features infrastructural stuff, because we don't do anything
interesting and just stub out the macro implementation for Optional.map,
but it's still very instructive. Note how the use of macro bundles
(a new feature in Scala 2.11) allows us to easily share common code between
macro implementations. This is something that's much harder to achieve in Scala 2.10.