step 28: doing it the hard way with typingTransform
If you stumble upon a refcheck problem, you're in for a treat!

When you need to insert untyped trees under typed trees and can't work
around that, the only robust way to proceed is to typecheck those untyped
trees, which involves manual creation of symbols and calling into
c.internal.typingTransform (of course, there's always c.untypecheck,
but as we discussed before it has potential to cause irreversible corruption of trees).

A typical course of action here is to take typed trees that
you plan to insert untyped fragments into and then call typingTransform
on them. typingTransform works very much like a normal Transformer, except
that, as it descends down the tree, it keeps track of lexical contexts
and then provides an API to typecheck trees in current context.

Your goal is to use the transformer to navigate to locations where you
want to insert untyped fragments and then at those locations use the
typechecking API before inserting the trees.

One thing that you'll need to remember is that the typechecker encapsulated
by typingTransform doesn't keep track of locally declared variables,
so you'll have to manually resolve references to those. If a referred variable
comes from a typed chunk of code, simply splice it into the quasiquotes
you generate. If a referred variable comes from untyped code, you'll have
to create a symbol for it first.

This is a somewhat gray area of the macro API (hence it's the part of
the `internal` submodule), so there isn't much documentation about it.
If you need more information, take a look at https://github.com/scala/async
for a great case study.