step 30: almost everything we've written can be ported to Scala 2.10
Macro bundles, relaxed rules for macro impl signatures and `showCode`
are available only in Scala 2.11, but their absense can be worked around.
It's annoying, but not fatal.

Quasiquotes, the most important highlight of Scala 2.11, are not shipping
with the official distribution of Scala 2.10, but there's the macro paradise
compiler plugin that provides this functionality.

What's even more interesting is that all the code that's been written can be
made to cross-compile against both Scala 2.10 and Scala 2.11. That requires
quite a bit of effort though - you can take a look at how this is done
in this commit.